% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check.R
\name{all_unique}
\alias{all_unique}
\title{Check if all values in a vector are unique}
\usage{
all_unique(x)
}
\arguments{
\item{x}{the vector to check}
}
\value{
TRUE if all values in the vector are unique
}
\description{
Check if all values in a vector are unique
}
\examples{
all_unique(c(1, 2, 3)) # TRUE
all_unique(c(1, 2, 2)) # FALSE
}
