% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_task_errors.R
\name{any_task_errors}
\alias{any_task_errors}
\alias{any_task_errors.benchmark_tbl}
\title{Check if any tasks produced errors}
\usage{
any_task_errors(x, verbose)

\method{any_task_errors}{benchmark_tbl}(x, verbose = FALSE)
}
\arguments{
\item{x}{the tibble to check}

\item{verbose}{TRUE if the rows with errors should be reported}
}
\value{
TRUE if any entry in the result column is a task_error object
}
\description{
Check the results column of a benchmark tibble for any task_error objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{any_task_errors(benchmark_tbl)}: 

}}
