% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_pipeline_list}
\alias{as_pipeline_list}
\title{convert to list of results with pipeline as name}
\usage{
as_pipeline_list(x)
}
\arguments{
\item{x}{the benchmark_tbl object to convert}
}
\description{
convert to list of results with pipeline as name
}
