% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellbench_casestudy.R
\name{cellbench_case_study}
\alias{cellbench_case_study}
\title{Open vignetted containing a case study using CellBench}
\usage{
cellbench_case_study()
}
\description{
Open vignetted containing a case study using CellBench
}
\examples{
\dontrun{
cellbench_case_study()
}
}
