% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clear_cached_datasets}
\alias{clear_cached_datasets}
\title{Clear cached datasets}
\usage{
clear_cached_datasets()
}
\description{
Delete the datasets cached by the load_*_data set of functions
}
\examples{
\dontrun{
clear_cached_datasets()
}
}
