% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{clear_cellbench_cache}
\alias{clear_cellbench_cache}
\title{Clear CellBench Cache}
\usage{
clear_cellbench_cache()
}
\description{
Clears the method cache for CellBench
}
\examples{
\dontrun{
clear_cellbench_cache()
}

}
