% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_clustering.R
\name{clustering_raceid}
\alias{clustering_raceid}
\title{Clustering using RaceID}
\usage{
clustering_raceid(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}
}
\description{
Clustering using RaceID
}
