% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_clustering.R
\name{clustering_sc3}
\alias{clustering_sc3}
\title{Clustering using SC3}
\usage{
clustering_sc3(sce, col.sym = "Symbol")
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}

\item{col.sym}{the column name containing gene symbols}
}
\description{
Clustering using SC3
}
