% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_clustering.R
\name{clustering_seurat}
\alias{clustering_seurat}
\title{Clustering using Seurat}
\usage{
clustering_seurat(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object to perform clustering on}
}
\description{
Clustering using Seurat
}
