% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compute_pca}
\alias{compute_pca}
\alias{compute_pca_most_var}
\title{Compute PCA coordinates}
\usage{
compute_pca(gene_expr, ndims = 2)

compute_pca_most_var(gene_expr, ndims = 2, ngenes = 500)
}
\arguments{
\item{gene_expr}{the gene expression matrix with samples in columns and expression in rows}

\item{ndims}{the number of dimensions to return}

\item{ngenes}{the number of most variable genes to use}
}
\value{
data.frame with PCA dimensions labelled Dim(n) where (n) is the dimension number
}
\description{
Compute PCA coordinates

Compute the PCA coordinates using the most variable genes
}
\section{Functions}{
\itemize{
\item \code{compute_pca_most_var}: 
}}

