% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{data_list}
\alias{data_list}
\title{Constructor for a data list}
\usage{
data_list(...)
}
\arguments{
\item{...}{objects, must all be named}
}
\value{
a list of named data
}
\description{
Constructor for a list of data, a thin wrapper around list() which
checks that all the inputs are of the same type and have names
}
\examples{
data(iris)
flist <- data_list(
    data1 = iris[1:20, ],
    data2 = iris[21:40, ]
)
}
