% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter.R
\name{filter_zero_genes}
\alias{filter_zero_genes}
\title{Filter out zero count genes}
\usage{
filter_zero_genes(x)
}
\arguments{
\item{x}{the SingleCellExperiment or matrix to filter}
}
\value{
object of same type as input with all zero count genes removed
}
\description{
Remove all genes (rows) where the total count is 0
}
\examples{
x <- matrix(rep(0:5, times = 5), nrow = 6, ncol = 5)
filter_zero_genes(x)
}
