% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_list.R
\name{fn_list}
\alias{fn_list}
\title{Constructor for a function list}
\usage{
fn_list(...)
}
\arguments{
\item{...}{objects, must all be named}
}
\value{
a list of named functions
}
\description{
Constructor for a list of functions, a thin wrapper around list() which
checks that all the inputs are functions and have names
}
\examples{
flist <- fn_list(
    mean = mean,
    median = median
)
}
