% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_impute.R
\name{impute_basics}
\alias{impute_basics}
\title{Impute using knn BaSICS}
\usage{
impute_basics(expr, ...)
}
\arguments{
\item{expr}{the expression matrix to impute on}

\item{...}{additional arguments for BaSICS}
}
\value{
matrix of imputed expression values
}
\description{
Impute using knn BaSICS
}
