% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_impute.R
\name{impute_drimpute}
\alias{impute_drimpute}
\title{Impute using DrImpute}
\usage{
impute_drimpute(expr)
}
\arguments{
\item{expr}{the expression matrix to impute on}
}
\value{
matrix of imputed expression values
}
\description{
Impute using DrImpute
}
