% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_impute.R
\name{impute_knn_smooth}
\alias{impute_knn_smooth}
\title{Impute using knn smooth}
\usage{
impute_knn_smooth(expr, k = ceiling(ncol(expr)/20))
}
\arguments{
\item{expr}{the expression matrix to impute on}

\item{k}{the number of nearest neighbours to use for smoothing}
}
\value{
matrix of imputed expression values
}
\description{
Impute using knn smooth
}
