% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_task_error.R
\name{is.task_error}
\alias{is.task_error}
\title{Check for task errors}
\usage{
is.task_error(x)
}
\arguments{
\item{x}{the object to be tested}
}
\value{
vector of logicals denoting if elements of the list are task_error objects
}
\description{
This is a helper function for checking the result column of a benchmark_tbl
for task_error objects. This is useful for filtering out rows where the
result is a task error.
}
