% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter.R
\name{keep_high_count_cells}
\alias{keep_high_count_cells}
\title{Filter down to the highest count cells}
\usage{
keep_high_count_cells(x, n)
}
\arguments{
\item{x}{the SingleCellExperiment or matrix}

\item{n}{the number of highest count cells to keep}
}
\value{
object of same type as input containing the highest count cells
}
\description{
Filter a SingleCellExperiment or matrix down to the cells (columns) with the
highest counts
}
\examples{
data(sample_sce_data)
keep_high_count_cells(sample_sce_data, 10)
}
