% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter.R
\name{keep_high_count_genes}
\alias{keep_high_count_genes}
\title{Filter down to the highest count genes}
\usage{
keep_high_count_genes(x, n)
}
\arguments{
\item{x}{the SingleCellExperiment or matrix}

\item{n}{the number of highest count genes to keep}
}
\value{
object of same type as input containing the highest count genes
}
\description{
Filter a SingleCellExperiment or matrix down to the genes (rows) with the
highest counts
}
\examples{
data(sample_sce_data)
keep_high_count_genes(sample_sce_data, 300)
}
