% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_filter.R
\name{keep_high_var_genes}
\alias{keep_high_var_genes}
\title{Filter down to the most variable genes}
\usage{
keep_high_var_genes(x, n)
}
\arguments{
\item{x}{the SingleCellExperiment or matrix}

\item{n}{the number of most variable genes to keep}
}
\value{
object of same type as input containing the most variable genes
}
\description{
Filter a SingleCellExperiment or matrix down to the most variable genes
(rows), variability is determined by var() scaled by the total counts for the
gene.
}
\examples{
data(sample_sce_data)
keep_high_var_genes(sample_sce_data, 50)
}
