% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_normalise.R
\name{lib_size_norm_expr}
\alias{lib_size_norm_expr}
\title{Library size normalise and transform to log scale}
\usage{
lib_size_norm_expr(x, log = TRUE, offset = 1)
}
\arguments{
\item{x}{the SingleCellExperiment or count matrix}

\item{log}{whether output should be log-scale}

\item{offset}{the counts to add before taking log}
}
\value{
matrix of normalised values
}
\description{
Library size normalise and transform to log scale
}
