% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_sc_data}
\alias{load_sc_data}
\alias{load_cell_mix_data}
\alias{load_mrna_mix_data}
\alias{load_all_data}
\title{Load CellBench Data}
\usage{
load_sc_data()

load_cell_mix_data()

load_mrna_mix_data()

load_all_data()
}
\value{
list of SingleCellExperiment
}
\description{
Load in all CellBench data described \href{https://github.com/LuyiTian/cellbench_file/blob/25f73cfcc87a84282cf55b4ac963032cc4da4988/README.md#summary-of-all-datasets}{here}.
}
\section{Functions}{
\itemize{
\item \code{load_sc_data}: Load single cell data

\item \code{load_cell_mix_data}: Load cell mixture data

\item \code{load_mrna_mix_data}: Load mrna mixture data
}}

\examples{
\dontrun{
cellbench_file <- load_all_data()
}
}
\keyword{internal}
