% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mhead}
\alias{mhead}
\title{Get head of 2 dimensional object as a square block}
\usage{
mhead(x, n = 6)
}
\arguments{
\item{x}{the object with 2 dimensions}

\item{n}{the size of the n-by-n block to extract}
}
\value{
an n-by-n sized subset of x
}
\description{
head prints all columns which may flood the console, mhead takes a square
block which can look nicer and still provide a good inspection of the
contents
}
\examples{
x <- matrix(runif(100), nrow = 10, ncol = 10)

mhead(x)
mhead(x, n = 3)
}
