% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_cells}
\alias{sample_cells}
\title{Sample cells from a SingleCellExperiment}
\usage{
sample_cells(x, n)
}
\arguments{
\item{x}{the SingleCellExperiment object}

\item{n}{the number of cells to sample}
}
\value{
SingleCellExperiment object
}
\description{
Sample n cells from a SingleCellExperiment object with no replacement.
}
\examples{
sample_sce_data <- readRDS(cellbench_file("celseq_sce_sample.rds"))
dim(sample_sce_data)
x <- sample_cells(sample_sce_data, 10)
dim(x)
}
