% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sampling.R
\name{sample_genes}
\alias{sample_genes}
\title{Sample genes from a SingleCellExperiment}
\usage{
sample_genes(x, n)
}
\arguments{
\item{x}{the SingleCellExperiment object}

\item{n}{the number of genes to sample}
}
\value{
SingleCellExperiment object
}
\description{
Sample n genes from a SingleCellExperiment object with no replacement
}
\examples{
data(sample_sce_data)
dim(sample_sce_data)
x <- sample_genes(sample_sce_data, 50)
dim(x)
}
