% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sce_data.R
\docType{data}
\name{sample_sce_data}
\alias{sample_sce_data}
\title{This is data for testing functions in CellBench.
It contains 200 genes and 50 cells randomly sampled from the CelSeq mRNA mixture dataset}
\format{An object of class \code{SingleCellExperiment} with 200 rows and 50 columns.}
\usage{
sample_sce_data
}
\description{
This is data for testing functions in CellBench.
It contains 200 genes and 50 cells randomly sampled from the CelSeq mRNA mixture dataset
}
\keyword{datasets}
