% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_normalise.R
\name{scran_norm_expr}
\alias{scran_norm_expr}
\title{Scran normalisation}
\usage{
scran_norm_expr(x, log = TRUE)
}
\arguments{
\item{x}{the SingleCellExperiment object}

\item{log}{TRUE if normalised values should be returned on log-scale}
}
\value{

}
\description{
Scran normalisation
}
