% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_cellbench_bpparam}
\alias{set_cellbench_bpparam}
\title{Set BiocParallel parameter used CellBench}
\usage{
set_cellbench_bpparam(param)
}
\arguments{
\item{param}{the BiocParallel parameter object}
}
\value{
None
}
\description{
This is a more advanced interface for changing CellBench's parallelism
settings. Internally CellBench uses BiocParallel for parallism, consult
the documentation of BiocParallel to see what settings are available.
}
\examples{
set_cellbench_threads(1) # CellBench runs on a single thread

}
\seealso{
\code{\link{set_cellbench_threads}} for more basic interface
}
