% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_cellbench_cache_path}
\alias{set_cellbench_cache_path}
\title{Set CellBench cache path}
\usage{
set_cellbench_cache_path(path = "./.CellBenchCache")
}
\arguments{
\item{path}{the path to where method caches should be stored}
}
\value{
None
}
\description{
Set CellBench cache path
}
\examples{
\dontrun{
set_cellbench_cache(".CellBenchCache")) # hidden folder in local path
}
set_cellbench_cache(file.path(tempdir(), ".CellBenchCache")) # store in temp directory valid for this session

}
\seealso{
[cache_method()]
}
