% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_cellbench_threads}
\alias{set_cellbench_threads}
\title{Set number of threads used by CellBench}
\usage{
set_cellbench_threads(nthreads = 1)
}
\arguments{
\item{nthreads}{the number of threads used by CellBench}
}
\value{
None
}
\description{
Sets global parameter for CellBench to use multiple threads for applying
methods. If any methods applied are multi-threaded then it's recommended to
set CellBench threads to 1. It only recommended to use CellBench with
multiple threads if methods applied can be set to run on single threads.
}
\examples{
set_cellbench_threads(1) # CellBench runs on a single thread

}
\seealso{
[set_cellbench_bpparam()] for more advanced interface
}
