% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_timing_tbl_methods.R
\name{strip_timing.benchmark_timing_tbl}
\alias{strip_timing.benchmark_timing_tbl}
\title{importFrom rlang .data}
\usage{
\method{strip_timing}{benchmark_timing_tbl}(x)
}
\description{
importFrom rlang .data
}
