% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_timing_tbl_methods.R
\name{strip_timing}
\alias{strip_timing}
\alias{strip_timing.benchmark_timing_tbl}
\title{Strip timing information}
\usage{
strip_timing(x)

\method{strip_timing}{benchmark_timing_tbl}(x)
}
\arguments{
\item{x}{the benchmark_timing_tbl object}
}
\value{
benchmark_tbl
}
\description{
Takes the result of a time_methods() call and remove timing information from
the `timed_result` column, replacing it with a `result` column and converting
it to a benchmark_tbl.
}
\examples{
\dontrun{
datasets <- list(
    data1 = 1:1e8,
)

transforms <- list(
    log = log,
    sqrt = sqrt
)

datasets \%>\%
    time_methods(transforms) \%>\%
    strip_timing()
}

}
