% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_tbl_methods.R
\name{summary.benchmark_tbl}
\alias{summary.benchmark_tbl}
\title{Summary of benchmark_tbl}
\usage{
\method{summary}{benchmark_tbl}(object, ...)
}
\arguments{
\item{object}{the benchmark_tbl to be summarised}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
None
}
\description{
Summary of benchmark_tbl
}
\examples{
# list of data
datasets <- list(
    set1 = rnorm(500, mean = 2, sd = 1),
    set2 = rnorm(500, mean = 1, sd = 2)
)

# list of functions
add_noise <- list(
    none = identity,
    add_bias = function(x) { x + 1 }
)

res <- apply_methods(datasets, add_noise)
summary(res)
}
\keyword{internal}
