% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_methods.R
\name{apply_methods}
\alias{apply_methods}
\title{Apply methods}
\usage{
apply_methods(x, fn_list, .name = NULL, suppress.messages = TRUE)
}
\arguments{
\item{x}{the list of data or benchmark tibble to apply methods to}

\item{fn_list}{the list of methods to be applied}

\item{.name}{(optional) the name of the column for methods applied}

\item{suppress.messages}{TRUE if messages from running methods should be
suppressed}
}
\value{
benchmark_tbl object containing results from methods applied
}
\description{
Apply methods
}
\examples{
# list of data
datasets <- list(
    set1 = rnorm(500, mean = 2, sd = 1),
    set2 = rnorm(500, mean = 1, sd = 2)
)

# list of functions
add_noise <- list(
    none = identity,
    add_bias = function(x) { x + 1 },
)

res <- apply_methods(datasets, add_noise)

}
