% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{cache_method}
\alias{cache_method}
\title{Create a cached function for CellBench}
\usage{
cache_method(f, cache = getOption("CellBench.cache"))
}
\arguments{
\item{f}{the function to be cached}

\item{cache}{the cache information (from memoise package)}
}
\value{
function object whose results are cached
}
\description{
Take a function and return a cached version. The arguments and results of a
cached method is saved in a file and if the cached function is called again
with the same arguments then the results will be retrieved from the cache
rather than be recomputed. \bold{(CAUTION)} Because cached functions called
with the same argument will always return the same output, pseudo-random
methods will not return varying results over repeated runs as one might
expect.
}
\examples{
set_cellbench_cache(file.path(tempdir(), ".CellBenchCache"))
cached_f <- cache_method(f)

}
