% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CellBench-package}
\alias{CellBench-package}
\alias{CellBench}
\title{A framework for benchmarking combinations of methods in multi-stage pipelines}
\description{
A framework for benchmarking combinations of methods in multi-stage pipelines
}
\seealso{
The core function in this package is \code{\link{apply_methods}}, see \code{vignette("Introduction", package = "CellBench")} for basic usage
}
\author{
Shian Su <\url{https://www.github.com/shians}>
}
