% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CellBench-package}
\alias{CellBench-package}
\alias{CellBench}
\title{A framework for benchmarking combinations of methods}
\description{
A framework for benchmarking combinations of methods
}
\seealso{
The core function in this package: \code{\link{apply_methods}}
}
\author{
Shian Su <\url{https://www.github.com/shians}>
}
