% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{cellbench_file}
\alias{cellbench_file}
\title{Get path to CellBench packaged data}
\usage{
cellbench_file(filename = NULL)
}
\arguments{
\item{filename}{the name of the file to look for}
}
\value{
string containing the path to the packaged data
}
\description{
Search CellBench package for packaged data, leaving argument empty will list
the available data.
}
