% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pipeline_collapse}
\alias{pipeline_collapse}
\title{Collapse benchmark_tbl into a two column summary}
\usage{
pipeline_collapse(x, sep = arrow_sep("right"), drop.steps = TRUE)
}
\arguments{
\item{x}{the benchmark_tbl to collapse}

\item{sep}{the separator to use for concatenating the pipeline steps}

\item{drop.steps}{if the data name and methods steps should be dropped from
the output. TRUE by default.}
}
\value{

}
\description{
Collapse benchmark_tbl into two columns: "pipeline" and "result". The
"pipeline" column will be the concatenated values from the data and methods
columns while the "result" column remains unchanged from the benchmark_tbl.
This is useful for having a string summary of the pipeline for annotating.
}
