% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fn_arg_seq}
\alias{print.fn_arg_seq}
\title{Print method for fn_arg_seq output}
\usage{
\method{print}{fn_arg_seq}(x, ...)
}
\arguments{
\item{x}{fn_arg_seq object}

\item{...}{addition arguments for print}
}
\value{
None
}
\description{
Print method for fn_arg_seq output
}
\examples{
fn_seq <- fn_arg_seq(kmeans, centers = 1:3)
fn_seq
}
\keyword{internal}
