% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sce_data.R
\docType{data}
\name{sample_sce_data}
\alias{sample_sce_data}
\title{This is data for testing functions in CellBench.}
\format{A SingleCellExperiment object with sample annotations in
  \code{colData(sample_sce_data)}. The annotation contains various QC metrics
  as well as the cell line mixture proportions
\describe{
  \item{H2228_prop}{proportion of mRNA from H2228 cell line}
  \item{H1975_prop}{proportion of mRNA from H1975 cell line}
  \item{HCC827_prop}{proportion of mRNA from HCC827 cell line}
}}
\usage{
data(sample_sce_data)
}
\description{
A dataset containing 200 genes and 50 cells randomly sampled from the CelSeq
mRNA mixture dataset, each sample is a mixture of mRNA material from 3
different human adenocarcinoma cell lines. Useful for quick prototyping of
method wrappers.
}
\seealso{
\code{\link{load_mrna_mix_data}}
}
\keyword{datasets}
