% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_cellbench_cache_path}
\alias{set_cellbench_cache_path}
\title{Set CellBench cache path}
\usage{
set_cellbench_cache_path(path = "./.CellBenchCache")
}
\arguments{
\item{path}{the path to where method caches should be stored}
}
\value{
None
}
\description{
Set CellBench cache path
}
\examples{
\dontrun{
# hidden folder in local path
set_cellbench_cache_path(".CellBenchCache"))
}
# store in temp directory valid for this session
set_cellbench_cache_path(file.path(tempdir(), ".CellBenchCache"))

}
\seealso{
\code{\link{cache_method}} for constructing cached methods.
}
