% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.benchmark_tbl}
\alias{summary.benchmark_tbl}
\title{Summary of benchmark_tbl}
\usage{
\method{summary}{benchmark_tbl}(x)
}
\arguments{
\item{x}{the benchmark_tbl to be summarised}
}
\description{
Summary of benchmark_tbl
}
\examples{
\dontrun{
res <- apply_methods(datasets, methods)
summary(res)
}
}
